Remove weaker key exchanges from ssh_config
Recommended by
https://stribika.github.io/2015/01/04/secure-secure-shell.html .

In a later commit, when I reformat this file, it will be evident that
my intention here is to prevent the user from *unknowingly*
connecting to servers that only support weak ciphers/key
exhanges/MACs. If the user really is okay will using a weak
component, then they can add an override to the file, otherwise they
may want to take additional precautions to secure themselves.